pyproject.toml
# File: pyproject.toml

[project]
name = "my_framework"
version = "1.0"
description = "Journalist with Reasoning Engine."
dependencies = [
    "pydantic>=2.0",
    "openai>=1.3",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "faiss-cpu>=1.7",
    "tiktoken>=0.5",
    "python-dotenv>=1.0",
    "newsapi-python>=0.2.7",
    "requests",
    "beautifulsoup4",
    "selenium>=4.0.0",
    "webdriver-manager>=4.0.0" # <--- ADD THIS LINE
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]